Delete final reference of REMOVED_TYPES, in V2 migration (#220685)
## Summary

Resolves https://github.com/elastic/kibana/issues/218843

This PR removes the final reference of `REMOVED_TYPES` that was in V2
migration algorithm. Now removed types is retrieved from the
typeRegistry as it should since
https://github.com/elastic/kibana/pull/218139. Thanks to this we can now
move `REMOVED_TYPES` to `core-saved-objects-server-internal` because the
circular dependency is not there anymore.

The ticket also mentioned:

```
We can also take advantage of this task and remove excludeOnUpgradeQuery ... We can remove it and create it when needed using the type registry legacy types, this way we don't need to carry that query string in all states.
```

But when doing this I noticed that this value is actually being updated
inside the migration so decided not to change it for now.

https://github.com/elastic/kibana/blob/7220f7a3ef610da1620b29e5f8fa9eae065135e0/src/core/packages/saved-objects/migration-server-internal/src/model/model.ts#L737-L754

### Checklist

Check the PR satisfies following conditions.

Reviewers should verify this PR satisfies this list as well.

- [x] [Unit or functional
tests](https://www.elastic.co/guide/en/kibana/master/development-tests.html)
were updated or added to match the most common scenarios
- [x] This was checked for breaking HTTP API changes, and any breaking
changes have been approved by the breaking-change committee. The
`release_note:breaking` label should be applied in these situations.
- [x] The PR description includes the appropriate Release Notes section,
and the correct `release_note:*` label is applied per the
[guidelines](https://www.elastic.co/guide/en/kibana/master/contributing.html#kibana-release-notes-process)

---------

(cherry picked from commit a5e8a0ad377ea0c98eefcddebdbe3e07bc1dca6c)

# Conflicts:
#	src/core/packages/saved-objects/base-server-internal/src/constants.ts
