+++ src/core/packages/saved-objects/base-server-internal/index.ts
@@ -7,7 +7,7 @@
  * License v3.0 only", or the "Server Side Public License, v 1".
  */
 
-export { DEFAULT_INDEX_TYPES_MAP, HASH_TO_VERSION_MAP, REMOVED_TYPES } from './src/constants';
+export { DEFAULT_INDEX_TYPES_MAP, HASH_TO_VERSION_MAP } from './src/constants';
 export { LEGACY_URL_ALIAS_TYPE, type LegacyUrlAlias } from './src/legacy_alias';
 export {
   getProperty,
