+++ src/core/packages/saved-objects/migration-server-internal/src/next.test.ts
@@ -21,7 +21,8 @@ describe('migrations v2 next', () => {
       (() => {}) as any,
       waitGroup(),
       waitGroup(),
-      waitGroup()
+      waitGroup(),
+      []
     )(state);
     expect(action).toEqual(null);
   });
@@ -32,7 +33,8 @@ describe('migrations v2 next', () => {
       (() => {}) as any,
       waitGroup(),
       waitGroup(),
-      waitGroup()
+      waitGroup(),
+      []
     )(state);
     expect(action).toEqual(null);
   });
