+++ src/core/packages/saved-objects/migration-server-internal/src/next.ts
@@ -12,7 +12,6 @@ import * as Option from 'fp-ts/lib/Option';
 import * as TaskEither from 'fp-ts/lib/TaskEither';
 import { omit } from 'lodash';
 import type { ElasticsearchClient } from '@kbn/core-elasticsearch-server';
-import { REMOVED_TYPES } from '@kbn/core-saved-objects-base-server-internal';
 import type { WaitGroup } from './kibana_migrator_utils';
 import type {
   AllActionStates,
@@ -77,7 +76,8 @@ export const nextActionMap = (
   transformRawDocs: TransformRawDocs,
   readyToReindex: WaitGroup<void>,
   doneReindexing: WaitGroup<void>,
-  updateRelocationAliases: WaitGroup<Actions.AliasAction[]>
+  updateRelocationAliases: WaitGroup<Actions.AliasAction[]>,
+  removedTypes: string[]
 ) => {
   return {
     INIT: (state: InitState) =>
@@ -104,7 +104,7 @@ export const nextActionMap = (
         excludeOnUpgradeQuery: state.excludeOnUpgradeQuery,
         excludeFromUpgradeFilterHooks: state.excludeFromUpgradeFilterHooks,
         knownTypes: state.knownTypes,
-        removedTypes: REMOVED_TYPES,
+        removedTypes,
       }),
     CLEANUP_UNKNOWN_AND_EXCLUDED_WAIT_FOR_TASK: (
       state: CleanupUnknownAndExcludedWaitForTaskState
@@ -322,14 +322,16 @@ export const next = (
   transformRawDocs: TransformRawDocs,
   readyToReindex: WaitGroup<void>,
   doneReindexing: WaitGroup<void>,
-  updateRelocationAliases: WaitGroup<Actions.AliasAction[]>
+  updateRelocationAliases: WaitGroup<Actions.AliasAction[]>,
+  removedTypes: string[]
 ) => {
   const map = nextActionMap(
     client,
     transformRawDocs,
     readyToReindex,
     doneReindexing,
-    updateRelocationAliases
+    updateRelocationAliases,
+    removedTypes
   );
   return (state: State) => {
     const delay = createDelayFn(state);
