+++ src/core/packages/saved-objects/migration-server-internal/src/run_resilient_migrator.ts
@@ -118,7 +118,14 @@ export async function runResilientMigrator({
   return migrationStateActionMachine({
     initialState,
     logger,
-    next: next(client, transformRawDocs, readyToReindex, doneReindexing, updateRelocationAliases),
+    next: next(
+      client,
+      transformRawDocs,
+      readyToReindex,
+      doneReindexing,
+      updateRelocationAliases,
+      typeRegistry.getLegacyTypes()
+    ),
     model,
     abort: async (state?: State) => {
       // At this point, we could reject this migrator's defers and unblock other migrators
